fix4: исправление замечаний GitHub Actions

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -17,5 +17,5 @@ public class Film {
     @Past
     private LocalDate releaseDate;
     @Positive
-    private double duration;
+    private int duration;
 }
